Add a version flag, update packages

+++ main.go
@@ -3,6 +3,7 @@ package main
 import (
 	"context"
 	"flag"
+	"fmt"
 	"os"
 	"os/signal"
 
@@ -10,7 +11,19 @@ import (
 	"github.com/maratig/trace_analyzer/internal/server"
 )
 
+var (
+	appName = "trace_analyzer"
+	version = "0.0.1"
+)
+
 func main() {
+	ver := flag.Bool("version", false, "print version and exit")
+	flag.Parse()
+	if ver != nil && *ver {
+		fmt.Println(appName, "version", version)
+		os.Exit(0)
+	}
+
 	ctx, _ := signal.NotifyContext(context.Background(), os.Kill, os.Interrupt)
 	port := flag.Int("port", 0, "port to be used in REST endpoint")
 	endpointConnectionWait := flag.Int(
